+++ packages/api/src/feature/meeting/announcement/announcement.module.ts
@@ -1,0 +1,10 @@
+import { Module } from "@nestjs/common";
+
+import { AnnouncementController } from "./announcement.controller";
+import { AnnouncementService } from "./announcement.service";
+
+@Module({
+  providers: [AnnouncementService],
+  controllers: [AnnouncementController],
+})
+export class AnnouncementModule {}
