+++ packages/api/src/feature/meeting/meeting.module.ts
@@ -4,13 +4,14 @@ import { DrizzleModule } from "src/drizzle/drizzle.module";
 
 import UserModule from "../user/user.module";
 
+import { AnnouncementModule } from "./announcement/announcement.module";
 import MeetingController from "./meeting.controller";
 import { MeetingGateway } from "./meeting.gateway";
 import { MeetingRepository } from "./meeting.repository";
 import { MeetingService } from "./meeting.service";
 
 @Module({
-  imports: [DrizzleModule, UserModule],
+  imports: [DrizzleModule, UserModule, AnnouncementModule],
   controllers: [MeetingController],
   providers: [MeetingGateway, MeetingService, MeetingRepository],
 })
